requirements.txt
PIL
matplotlib
matplotlib.pyplot
numpy
scipy.ndimage
skimage.measure
PySide2